backups/backup_completo_20250912_005219/requirements.txt
# Django Core
Django==5.1.5
asgiref==3.8.1
sqlparse==0.5.3

# Django Extensions and Themes
django-jazzmin==3.0.1
django-crispy-forms==2.3
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10

# Data Analysis and Visualization
numpy==2.2.5
pandas==2.2.3
plotly==6.0.0
django-plotly-dash==2.4.5
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
django-chartjs==2.3.0
dpd_components==0.1.0

# Image Processing
Pillow==11.1.0

# Excel Support
openpyxl==3.1.5
et_xmlfile==2.0.0

# Web Framework (for dash)
Flask==3.0.3
Werkzeug==3.0.6
Jinja2==3.1.5
MarkupSafe==3.0.2
itsdangerous==2.2.0

# HTTP and Networking
requests==2.32.3
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# Date and Time
python-dateutil==2.9.0.post0
pytz==2025.1
tzdata==2025.1

# Utilities
click==8.1.8
colorama==0.4.6
blinker==1.9.0
six==1.17.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
retrying==1.3.4
tomli==2.2.1
typing_extensions==4.12.2
zipp==3.21.0
importlib_metadata==8.6.1
mypy_extensions==1.1.0
narwhals==1.26.0
nest-asyncio==1.6.0

# Development Tools
black==25.1.0
git-filter-repo==2.47.0

# WebSocket Support
channels==4.2.0

# PostgreSQL Support
psycopg2-binary==2.9.10
python-dotenv==1.1.1

# PostgreSQL Support
psycopg2-binary==2.9.10
python-dotenv==1.1.1
